64bit build warnings fixed
git-svn-id: https://svn.oktetlabs.ru/svnroot/prj/oktetlabs/test-environment/src/trunk@8197 9d14d920-75d5-0310-8bd7-a52572981aee

+++ lib/tad_ch/tad.h
@@ -258,7 +258,7 @@ typedef struct csap_instance
     struct timeval  last_pkt;    /**< moment of last good packet processed:
                                        matched in RX mode or sent in TX mode. */
 
-    size_t          num_packets; /**< number of good packets to be processed. */
+    unsigned int    num_packets; /**< number of good packets to be processed. */
     size_t          total_bytes; /**< quantity of total processed bytes in 
                                       last operation, for some protocols it is
                                       not sensible, but it is more convenient
